* origin/released:
  xserver xkeyboard-config plink mesa git update Jan 15 2019
  xserver mesalib plink git update Jan 10 2019
  xkeyboard-config libxcb mkfontscale plink mesa git update Jan 8 2019
  xkeyboard-config plink mesa git update Jan 4 2019
  xserver libX11 mesa plink xkbcomp git update Jan 2 2019

+++ libX11/src/xlibi18n/lcCharSet.c
@@ -172,7 +172,7 @@ _XlcCreateDefaultCharSet(
     const char *ct_sequence)
 {
     XlcCharSet charset;
-    int name_len, ct_sequence_len;
+    size_t name_len, ct_sequence_len;
     const char *colon;
     char *tmp;
 
@@ -195,7 +195,7 @@ _XlcCreateDefaultCharSet(
 
     /* Fill in encoding_name and xrm_encoding_name.  */
     if ((colon = strchr(charset->name, ':')) != NULL) {
-        unsigned int length = colon - charset->name;
+        size_t length = (size_t)(colon - charset->name);
         char *encoding_tmp = Xmalloc(length + 1);
         if (encoding_tmp == NULL) {
             Xfree((char *) charset->name);
